+++ examples/03_comparison/src/CalcFib.tsx
@@ -2,8 +2,10 @@ import * as React from 'react';
 
 import { useWorker } from 'react-hooks-worker';
 
+const worker = new Worker('./slow_fib', { type: 'module' });
+
 const CalcFib: React.FC<{ count: number }> = ({ count }) => {
-  const { result, error } = useWorker('./slow_fib.js', count);
+  const { result, error } = useWorker(worker, count);
   if (error) return <div>Error: {error}</div>;
   return <div>Result: {result}</div>;
 };
