+++ examples/03_comparison/src/slow_fib.js
@@ -1,0 +1,5 @@
+import { exposeWorker } from 'react-hooks-worker';
+
+const fib = i => (i <= 1 ? i : fib(i - 1) + fib(i - 2));
+
+exposeWorker(fib);
